Extracted dat methods into own middleware.
